+++ src/cli/mod.rs
@@ -1,5 +1,7 @@
 mod args;
 
+pub use args::Args;
+
 use crate::{
     core::OpenCloset,
     error::SafeClosetError,
@@ -10,18 +12,21 @@ use crate::{
 ///
 /// Starts the TUI if a path to a closet is given
 pub fn run() -> Result<(), SafeClosetError> {
-    let args: args::Args = argh::from_env();
+    let args: Args = argh::from_env();
     if args.version {
         println!("SafeCloset {}", env!("CARGO_PKG_VERSION"),);
         return Ok(());
     }
     info!("args: {:#?}", &args);
 
-    if let Some(path) = args.path {
-        let closet = OpenCloset::open_or_create(path)?;
-        tui::run(closet, args.hide)?;
+    if let Some(path) = &args.path {
+        let closet = OpenCloset::open_or_create(path.clone())?;
+        tui::run(closet, &args)?;
     } else {
-        println!("Please provide as argument the path to the closet file to create or open");
+        println!(
+            "Please provide as argument the path to the closet file to create or open, \
+            or use --help for help."
+        );
     }
 
     Ok(())